// generated-api/models/PredictionHistoryResponse.ts
export interface PredictionHistoryResponse {
    /**
     * Unique prediction ID
     * @type {number}
     * @memberof PredictionHistoryResponse
     */
    predictionId?: number;
    /**
     * Name of boxer A
     * @type {string}
     * @memberof PredictionHistoryResponse
     */
    boxerAName?: string;
    /**
     * Name of boxer B
     * @type {string}
     * @memberof PredictionHistoryResponse
     */
    boxerBName?: string;
    /**
     * Model-predicted winner
     * @type {string}
     * @memberof PredictionHistoryResponse
     */
    predictedWinner?: string;
    /**
     * Actual winner after the fight, if known
     * @type {string}
     * @memberof PredictionHistoryResponse
     */
    matchWinner?: string;
    /**
     * Actual method of victory, if known
     * @type {string}
     * @memberof PredictionHistoryResponse
     */
    matchWinMethod?: string;
    /**
     * Weight class ID
     * @type {number}
     * @memberof PredictionHistoryResponse
     */
    weightClassId?: number;
    /**
     * Closeness score of boxer A to the perfect boxer
     * @type {number}
     * @memberof PredictionHistoryResponse
     */
    boxerAClosenessScore?: number;
    /**
     * Closeness score of boxer B to the perfect boxer
     * @type {number}
     * @memberof PredictionHistoryResponse
     */
    boxerBClosenessScore?: number;
    /**
     * Predicted win probability for boxer A
     * @type {number}
     * @memberof PredictionHistoryResponse
     */
    probabilityA?: number;
    /**
     * Predicted win probability for boxer B
     * @type {number}
     * @memberof PredictionHistoryResponse
     */
    probabilityB?: number;
    /**
     * Detailed structured breakdown or explanation of the prediction
     * @type {{ [key: string]: object; }}
     * @memberof PredictionHistoryResponse
     */
    breakdownSnapshot?: { [key: string]: object; };
    /**
     * Date and time of prediction
     * @type {Date}
     * @memberof PredictionHistoryResponse
     */
    predictionDate?: Date;
}